# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Franchise Generation API
  description: |
    REST API for the Franchise Generation platform — a multi-brand franchise ecosystem
    connecting brands, candidates, investors, consultants, and real estate developers.

    ## Authentication
    All protected endpoints require a JWT Bearer token obtained from `POST /auth/login`.
    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```

    ## Base URL
    Production: `https://api.franchisegeneration.be/api/v1`
    Local dev: `http://localhost:3000/api/v1`

    ## Roles
    | Role | Description |
    |------|-------------|
    | `admin` | Full backoffice access |
    | `investor` | Investment portal access |
    | `candidate` | Candidate onboarding portal |
    | `consultant` | Lead management access |
    | `developer` | Real estate developer access |
    | `brand` | Brand management portal |

  version: 1.0.0
  contact:
    name: Franchise Generation
    url: https://franchisegeneration.be
servers:
  - url: /api/v1
    description: Current version
security:
  - BearerAuth: []
...
paths:
...
  /investors:
    get:
      tags: [Investors]
      summary: List investors
      description: Returns all investor profiles. Accessible by admin only.
      parameters:
        - name: tier
          in: query
          schema:
            type: string
            enum: [Bronze, Argent, Or, Privilège]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of investors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedList'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Investor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /investors/{id}:
    get:
      tags: [Investors]
      summary: Get investor by ID
      description: Returns the full investor profile. Investors can only access their own record.
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Investor detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Investor'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /investors/{id}/portfolio:
    get:
      tags: [Investors]
      summary: Get investor portfolio
      description: Returns all investments grouped by brand, including shop-level KPIs and performance charts.
      parameters:
        - $ref: '#/components/parameters/id'
        - name: brand
          in: query
          description: Filter to a specific brand
          schema:
            type: string
      responses:
        '200':
          description: Portfolio data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestorPortfolio'
              example:
                summary:
                  totalInvested: 85000
                  totalRepaid: 8580
                  totalValuation: 91200
                  roiTarget: 8.4
                  roiCurrent: 8.0
                brands:
                  - brandId: atelier
                    summary:
                      invested: 45000
                      repaid: 6200
                      valuation: 49600
                    shops:
                      - id: couq-chatelain
                        name: "Couq Châtelain"
                        city: Bruxelles
                        opened: "Juin 2024"
                        invested: 25000
                        repaid: 4100
                        health: good
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /investors/{id}/documents:
    get:
      tags: [Investors]
      summary: Get investor documents
      description: Returns all documents in the investor's document vault.
      parameters:
        - $ref: '#/components/parameters/id'
        - name: type
          in: query
          schema:
            type: string
            enum: [investment, ownership, kyc, aml, banking, tax, contract, schedule, statement]
        - name: brand
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [validated, pending, expired, missing, signed, review]
      responses:
        '200':
          description: Array of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /investors/{id}/repayments:
    get:
      tags: [Investors]
      summary: Get repayment schedule
      description: Returns all repayment schedule rows across all investments for the investor.
      parameters:
        - $ref: '#/components/parameters/id'
        - name: brand
          in: query
          schema:
            type: string
        - name: paid
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Array of repayment records
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    investmentId:
                      type: integer
                    shopName:
                      type: string
                    brandId:
                      type: string
                    dueDate:
                      type: string
                      format: date
                    amount:
                      type: number
                    paid:
                      type: boolean
                    paidAt:
                      type: string
                      format: date-time
                      nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
  /investors/{id}/interest:
    post:
      tags: [Investors]
      summary: Express investment interest
      description: Registers an investor's interest in a specific investment opportunity.
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [opportunityId, amount]
              properties:
                opportunityId:
                  type: string
                  example: mania-namur
                amount:
                  type: integer
                  description: Investment amount in EUR
                  example: 10000
                message:
                  type: string
                  example: "Je souhaite m'engager sur 10 000 € pour Mania Namur."
      responses:
        '201':
          description: Interest registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  ticketId:
                    type: string
                    description: Support ticket created for follow-up
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =========================================================================
  # LEADS
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Request validation failed.
            details:
              - field: email
                message: Must be a valid email address.
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: A valid Bearer token is required.
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: You do not have permission to access this resource.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: The requested resource was not found.
...
  schemas:
...
    Shop:
      type: object
      description: A physical franchise shop location
      properties:
        id:
          type: string
          example: couq-chatelain
        brandId:
          type: string
          example: couq
        name:
          type: string
          example: "Couq Châtelain"
        city:
          type: string
          example: Bruxelles
        kind:
          type: string
          example: Boulangerie
        regionId:
          type: string
          example: bxl
        address:
          type: string
          nullable: true
        surfaceM2:
          type: integer
          nullable: true
        opened:
          type: string
          example: "Juin 2024"
        status:
          type: string
          enum: [open, pre-opening, closed, paused]
        health:
          type: string
          enum: [good, at-risk, critical]
          example: good
        photoUrl:
          type: string
          format: uri
          nullable: true
        kpiSnapshot:
          type: object
          properties:
            ca:
              type: integer
              description: Monthly revenue in EUR
            profit:
              type: number
              description: Net profit margin percentage
            cust:
              type: integer
              description: Average customers per day
            basket:
              type: number
              description: Average basket in EUR
            vsBudget:
              type: number
              description: Revenue vs budget percentage
        consultant:
          type: object
          nullable: true
          properties:
            name:
              type: string
            lastVisit:
              type: string
            verdict:
              type: string
              enum: [on-track, at-risk, critical]
            summary:
              type: string
            action:
              type: string
        franchisee:
          type: object
          nullable: true
          properties:
            name:
              type: string
            tier:
              type: string
            overall:
              type: number
            since:
              type: string

    # -------------------------------------------------------------------------
...
    Investor:
      type: object
      description: An investor profile
      properties:
        id:
          type: string
          example: inv_001
        userId:
          type: string
        name:
          type: string
          example: Claire Vermeulen
        initials:
          type: string
          example: CV
        email:
          type: string
          format: email
          example: claire.vermeulen@example.com
        phone:
          type: string
          example: "+32 477 12 34 56"
        address:
          type: string
          example: "Avenue Louise 142 · 1050 Bruxelles"
        nationality:
          type: string
          example: Belge
        iban:
          type: string
          example: "BE68 5390 0754 7034"
        role:
          type: string
          example: "Investisseur particulier"
        tier:
          type: string
          enum: [Bronze, Argent, Or, Privilège]
          example: Privilège
        since:
          type: string
          example: "Juin 2023"
        totalInvested:
          type: number
        totalValuation:
          type: number
        totalPaid:
          type: number
        remaining:
          type: number

    # -------------------------------------------------------------------------
    InvestorPortfolio:
      type: object
      description: Full investor portfolio grouped by brand
      properties:
        summary:
          type: object
          properties:
            totalInvested:
              type: number
              example: 85000
            totalRepaid:
              type: number
              example: 8580
            totalValuation:
              type: number
              example: 91200
            roiTarget:
              type: number
              example: 8.4
            roiCurrent:
              type: number
              example: 8.0
        brands:
          type: array
          items:
            type: object
            properties:
              brandId:
                type: string
              summary:
                type: object
                properties:
                  invested:
                    type: number
                  repaid:
                    type: number
                  valuation:
                    type: number
                  roiTarget:
                    type: number
                  roiCurrent:
                    type: number
              ca12m:
                type: array
                items:
                  type: integer
                description: Monthly revenue series for the past 12 months
              budget12m:
                type: array
                items:
                  type: integer
                description: Monthly budget series for the past 12 months
              shops:
                type: array
                items:
                  $ref: '#/components/schemas/Shop'

    # -------------------------------------------------------------------------
...
    Document:
      type: object
      description: A document in the platform vault (contract, KYC, schedule, etc.)
      properties:
        id:
          type: string
          example: d-kyc-id
        ownerId:
          type: string
          description: ID of the investor or candidate who owns this document
        ownerType:
          type: string
          enum: [investor, candidate]
        brandId:
          type: string
          nullable: true
          example: atelier
        type:
          type: string
          enum: [investment, ownership, kyc, aml, banking, tax, contract, schedule, statement]
          example: kyc
        status:
          type: string
          enum: [validated, pending, expired, missing, signed, review]
          example: validated
        title:
          type: string
          example: "Pièce d'identité — Carte d'identité belge"
        sub:
          type: string
          example: "Recto + verso"
        filename:
          type: string
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
          description: Signed download URL (time-limited)
        size:
          type: string
          nullable: true
          example: "2,4 MB"
        date:
          type: string
          description: Document date (formatted)
          example: "15/06/2023"
        expiresAt:
          type: string
          format: date
          nullable: true
          example: "2033-06-15"
        createdAt:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Standard error response envelope
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error description
          example: Request validation failed.
        details:
          type: array
          nullable: true
          description: Field-level validation errors
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # -------------------------------------------------------------------------
    PaginatedList:
      type: object
      description: Generic paginated list wrapper
      properties:
        data:
          type: array
          items: {}
          description: Array of result items
        total:
          type: integer
          description: Total number of matching records
          example: 42
        page:
          type: integer
          description: Current page number (1-indexed)
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3